Changed floats to doubles

+++ src/main.cpp
@@ -23,8 +23,8 @@ void find_burn_in_time(int N, int L=20, double T=1){
     vector<double> expected_E_burn_in;
     vector<double> expected_M_burn_in;
     for (int burn_in=0; burn_in<N; burn_in+=max(1, N/100)){
-        map<int, float> buckets_E = stat_utils::distribution(sampled_E + burn_in, N - burn_in);
-        map<int, float> buckets_M = stat_utils::distribution(sampled_M + burn_in, N - burn_in);
+        map<int, double> buckets_E = stat_utils::distribution(sampled_E + burn_in, N - burn_in);
+        map<int, double> buckets_M = stat_utils::distribution(sampled_M + burn_in, N - burn_in);
         expected_E_burn_in.push_back(stat_utils::expected_value(buckets_E));
         expected_M_burn_in.push_back(stat_utils::expected_value(buckets_M));
         burn_in_time.push_back(burn_in);
@@ -61,18 +61,18 @@ void produce_distributions(const int iters, const int L, double T, int burn_in_t
         sampled_magnetization_abs[i] = abs(model.get_magnetization());
     }
     auto scale = [N](int x){return x / N;};
-    map<int, float> buckets_epsilon = stat_utils::distribution(sampled_energy, iters, scale);
-    map<int, float> buckets_m_abs = stat_utils::distribution(sampled_magnetization_abs, iters, scale);
+    map<int, double> buckets_epsilon = stat_utils::distribution(sampled_energy, iters, scale);
+    map<int, double> buckets_m_abs = stat_utils::distribution(sampled_magnetization_abs, iters, scale);
 
     auto square = [](int x){return x * x;};
-    float expected_epsilon = stat_utils::expected_value(buckets_epsilon);
-    float expected_m_abs = stat_utils::expected_value(buckets_m_abs);
-    float expected_energy = stat_utils::expected_value(sampled_energy, iters);
-    float expected_energy_sq = stat_utils::expected_value(sampled_energy, iters, square);
-    float expected_magnetization_abs = stat_utils::expected_value(sampled_magnetization_abs, iters);
-    float expected_magnetization_sq = stat_utils::expected_value(sampled_magnetization_abs, iters, square);
-    float c_v = (1. / N) * (1. / (T * T)) * (expected_energy_sq - expected_energy * expected_energy);
-    float chi = (1. / N) * (1. / T) * (expected_magnetization_sq - expected_magnetization_abs * expected_magnetization_abs);
+    double expected_epsilon = stat_utils::expected_value(buckets_epsilon);
+    double expected_m_abs = stat_utils::expected_value(buckets_m_abs);
+    double expected_energy = stat_utils::expected_value(sampled_energy, iters);
+    double expected_energy_sq = stat_utils::expected_value(sampled_energy, iters, square);
+    double expected_magnetization_abs = stat_utils::expected_value(sampled_magnetization_abs, iters);
+    double expected_magnetization_sq = stat_utils::expected_value(sampled_magnetization_abs, iters, square);
+    double c_v = (1. / N) * (1. / (T * T)) * (expected_energy_sq - expected_energy * expected_energy);
+    double chi = (1. / N) * (1. / T) * (expected_magnetization_sq - expected_magnetization_abs * expected_magnetization_abs);
 
     cout << "Sampled distribution of epsilon" << endl;
     stat_utils::print_map(buckets_epsilon);
